Rename ResultPage file
