+++ lib/result_page.dart
@@ -1,15 +1,44 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_redux/flutter_redux.dart';
-import 'package:vasculink/state_manager.dart';
-import 'package:vasculink/vasculink_app_bar.dart';
 
-class ResultPage extends StatelessWidget {
-  ResultPage();
+void main() => runApp(MyApp());
 
-  String getImage(int riskLevel) {
-    if (riskLevel < 3) {
+class MyApp extends StatelessWidget {
+  // This widget is the root of your application.
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Vasculink App',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+        visualDensity: VisualDensity.adaptivePlatformDensity,
+      ),
+      home: MyHomePage(0 ),// TODO: change 0 to the actual risk level value),
+    );
+  }
+}
+
+class MyHomePage extends StatefulWidget {
+
+  var _riskLevel;
+
+  MyHomePage(this._riskLevel);
+
+  @override
+  _MyHomePageState createState() => _MyHomePageState(_riskLevel);
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+
+  var imageName = 'images/HighRisk.png';
+
+  _MyHomePageState(int _riskLevel) {
+    this.imageName = getImage(_riskLevel);
+  }
+
+  String getImage(int _riskLevel) {
+    if (_riskLevel < 3) {
       return 'images/LowRisk.png';
-    } else if (riskLevel < 5) {
+    } else if (_riskLevel < 5) {
       return 'images/MediumRisk.png';
     }
     return 'images/HighRisk.png';
@@ -17,101 +46,119 @@ class ResultPage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: VasculinkAppBar().build(context),
-      backgroundColor: Colors.grey[100],
-      body: StoreConnector<List<RiskFactor>, int>(converter: (store) {
-        return store.state.fold(
-            0,
-            (runningSum, riskFactor) =>
-                runningSum + (riskFactor.value ? riskFactor.weight : 0));
-      }, builder: (context, riskLevel) {
-        // build the string for the appropriate risk level
-        String riskLevelText;
-        if (riskLevel < 3) {
-          riskLevelText = 'Low';
-        } else if (riskLevel < 5) {
-          riskLevelText = 'Medium';
-        } else {
-          riskLevelText = 'High';
-        }
 
-        return Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(
-              'Patient Score',
-              style: TextStyle(
-                color: Colors.blue[400],
-                fontSize: 28,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-            SizedBox(
-              height: 10,
-            ),
-            Card(
-              elevation: 0.0,
-              margin: EdgeInsets.symmetric(vertical: 10.0, horizontal: 60.0),
-              child: Column(
-                children: <Widget>[
-                  SizedBox(
-                    height: 20,
-                  ),
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: <Widget>[
-                      Text(
-                        riskLevel.toString(),
-                        style: TextStyle(
-                          color: Colors.blue[400],
-                          fontSize: 55,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                      Text(
-                        '/',
-                        style: TextStyle(
-                          color: Colors.blue[400],
-                          fontSize: 25,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                      Text(
-                        '7',
-                        style: TextStyle(
-                          color: Colors.blue[400],
-                          fontSize: 25,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                    ],
-                  ),
-                  SizedBox(
-                    height: 10,
+    return Scaffold(
+
+      appBar: PreferredSize(
+          preferredSize: Size.fromHeight(130.0),
+          child: AppBar(
+            elevation: 0,
+            flexibleSpace: Padding(
+              padding: const EdgeInsets.fromLTRB(0,65,0,0),
+              child: Row(
+                children:<Widget> [
+                  IconButton(
+                    icon: Icon(Icons.clear),
+                    tooltip: "cancel the entry",
+                    onPressed: (){
+                      // TODO: implement cancel function
+                    },
+                    color: Colors.white,
+                    iconSize: 30,
                   ),
                   Text(
-                    '$riskLevelText Risk',
+                    'Result',
                     style: TextStyle(
-                      color: Colors.black26,
-                      fontSize: 25,
+                      color: Colors.white,
+                      fontSize: 24,
                     ),
-                  ),
-                  SizedBox(
-                    height: 20,
                   ),
                 ],
               ),
             ),
-            Container(
-              child: Image.asset(
-                getImage(riskLevel),
-                height: 180,
+            backgroundColor: Colors.blue[400],
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.vertical(
+                bottom: Radius.circular(29),
               ),
             ),
-          ],
-        );
-      }),
+          )
+      ),
+
+      backgroundColor: Colors.grey[100],
+      body: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children:<Widget> [
+          Text(
+            'Patient Score',
+            style: TextStyle(
+              color: Colors.blue[400],
+              fontSize: 28,
+              fontWeight: FontWeight.bold,
+            ),
+          ),
+          SizedBox(
+            height: 10,
+          ),
+          Card(
+            elevation: 0.0,
+            margin: EdgeInsets.symmetric(vertical: 10.0, horizontal: 60.0),
+            child: Column(
+              children:<Widget> [
+                SizedBox(
+                  height: 20,
+                ),
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children:<Widget> [
+                    Text('4',
+                      style: TextStyle(
+                        color: Colors.blue[400],
+                        fontSize: 55,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                    Text('/',
+                      style: TextStyle(
+                        color: Colors.blue[400],
+                        fontSize: 25,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                    Text('7',
+                      style: TextStyle(
+                        color: Colors.blue[400],
+                        fontSize: 25,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                  ],
+                ),
+                SizedBox(
+                  height: 10,
+                ),
+                Text('Medium Risk',
+                  style: TextStyle(
+                    color: Colors.black26,
+                    fontSize: 25,
+                  ),
+                ),
+                SizedBox(
+                  height: 20,
+                ),
+              ],
+            ),
+          ),
+          Container(
+            child: Image.asset(
+              imageName,
+              height: 180,
+//                  width: 100,
+            ),
+          ),
+        ],
+      ),
+
       floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
       floatingActionButton: FloatingActionButton(
         backgroundColor: Colors.white,
